# Optimize Semaphore implementation