--- a/core-tests/shared/src/test/scala/zio/SemaphoreBenchmarkSpec.scala
+++ b/core-tests/shared/src/test/scala/zio/SemaphoreBenchmarkSpec.scala
@@ -1,0 +1,1 @@
+// This file has been moved to benchmarks/src/test/scala/zio/SemaphoreBenchmarkSpec.scala