--- a/project/Dependencies.scala
+++ b/project/Dependencies.scala
@@ -14,6 +14,7 @@ object Dependencies {
   val ScalaNativeCryptoVersion     = "0.2.1"
   val ScalaSecureRandomVersion     = "1.0.0"
   val ScalaJsDomVersion            = "2.8.1"
+  val JctoolsVersion               = "4.0.5"
 
   // Documentations and example dependencies
   val CatsEffectVersion = "3.6.3"