/claim #9093

This change introduces a new Semaphore implementation that is optimized for performance. It uses a custom WaiterQueue and a custom Waiter object to reduce contention and improve throughput.

Here are the demo videos. I had ti compress them because they were a bit too large
BENCHMARK RUN VIDEO:   
[BenchMark Run.zip](https://github.com/user-attachments/files/21528324/BenchMark.Run.zip)

BENCHMARK TESTS RUN VIDEO:   
[Test Run.zip](https://github.com/user-attachments/files/21528326/Test.Run.zip)

---------------------------------------------------------**BENCHMARK RESULTS**------------------------------------------------------
[info] # JMH version: 1.37
[info] # VM version: JDK 24.0.2, Java HotSpot(TM) 64-Bit Server VM, 24.0.2+12-54
[info] # VM invoker: C:\Program Files\Java\jdk-24\bin\java.exe
[info] # VM options: <none>
[info] # Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
[info] # Warmup: 3 iterations, 1 s each
[info] # Measurement: 3 iterations, 1 s each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Throughput, ops/time
[info] # Benchmark: zio.stm.SinglePermitSemaphoreBenchmark.javaSemaphoreFair
[info] # Parameters: (fibers = 1, permits = 1)
[info] # Run progress: 0.00% complete, ETA 00:00:48
[info] # Fork: 1 of 1
[info] WARNING: A terminally deprecated method in sun.misc.Unsafe has been called
[info] WARNING: sun.misc.Unsafe::objectFieldOffset has been called by org.openjdk.jmh.util.Utils (file:/C:/Users/USER/3D%20Objects/zio/target/bg-jobs/sbt_9d8cd5a8/target/309f7c90/1296d6b9/jmh-core-1.37.jar)
[info] WARNING: Please consider reporting this to the maintainers of class org.openjdk.jmh.util.Utils
[info] WARNING: sun.misc.Unsafe::objectFieldOffset will be removed in a future release
[info] # Warmup Iteration   1: 129.149 ops/s
[info] # Warmup Iteration   2: 2431.575 ops/s
[info] # Warmup Iteration   3: 4032.638 ops/s
[info] Iteration   1: 4985.660 ops/s
[info] Iteration   2: 4781.877 ops/s
[info] Iteration   3: 6809.672 ops/s
[info] Result "zio.stm.SinglePermitSemaphoreBenchmark.javaSemaphoreFair":
[info]   5525.736 ∩┐╜(99.9%) 20370.567 ops/s [Average]
[info]   (min, avg, max) = (4781.877, 5525.736, 6809.672), stdev = 1116.580
[info]   CI (99.9%): [? 0, 25896.304] (assumes normal distribution)
[info] # JMH version: 1.37
[info] # VM version: JDK 24.0.2, Java HotSpot(TM) 64-Bit Server VM, 24.0.2+12-54
[info] # VM invoker: C:\Program Files\Java\jdk-24\bin\java.exe
[info] # VM options: <none>
[info] # Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
[info] # Warmup: 3 iterations, 1 s each
[info] # Measurement: 3 iterations, 1 s each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Throughput, ops/time
[info] # Benchmark: zio.stm.SinglePermitSemaphoreBenchmark.javaSemaphoreFair
[info] # Parameters: (fibers = 2, permits = 1)
[info] # Run progress: 12.50% complete, ETA 00:00:50
[info] # Fork: 1 of 1
[info] # Warmup Iteration   1: WARNING: A terminally deprecated method in sun.misc.Unsafe has been called
[info] WARNING: sun.misc.Unsafe::objectFieldOffset has been called by org.openjdk.jmh.util.Utils (file:/C:/Users/USER/3D%20Objects/zio/target/bg-jobs/sbt_9d8cd5a8/target/309f7c90/1296d6b9/jmh-core-1.37.jar)
[info] WARNING: Please consider reporting this to the maintainers of class org.openjdk.jmh.util.Utils
[info] WARNING: sun.misc.Unsafe::objectFieldOffset will be removed in a future release
[info] 10.760 ops/s
[info] # Warmup Iteration   2: 422.371 ops/s
[info] # Warmup Iteration   3: 218.079 ops/s
[info] Iteration   1: 234.880 ops/s
[info] Iteration   2: 226.681 ops/s
[info] Iteration   3: 276.869 ops/s
[info] Result "zio.stm.SinglePermitSemaphoreBenchmark.javaSemaphoreFair":
[info]   246.143 ∩┐╜(99.9%) 491.175 ops/s [Average]
[info]   (min, avg, max) = (226.681, 246.143, 276.869), stdev = 26.923
[info]   CI (99.9%): [? 0, 737.318] (assumes normal distribution)
[info] # JMH version: 1.37
[info] # VM version: JDK 24.0.2, Java HotSpot(TM) 64-Bit Server VM, 24.0.2+12-54
[info] # VM invoker: C:\Program Files\Java\jdk-24\bin\java.exe
[info] # VM options: <none>
[info] # Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
[info] # Warmup: 3 iterations, 1 s each
[info] # Measurement: 3 iterations, 1 s each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Throughput, ops/time
[info] # Benchmark: zio.stm.SinglePermitSemaphoreBenchmark.javaSemaphoreUnfair
[info] # Parameters: (fibers = 1, permits = 1)
[info] # Run progress: 25.00% complete, ETA 00:00:43
[info] # Fork: 1 of 1
[info] WARNING: A terminally deprecated method in sun.misc.Unsafe has been called
[info] WARNING: sun.misc.Unsafe::objectFieldOffset has been called by org.openjdk.jmh.util.Utils (file:/C:/Users/USER/3D%20Objects/zio/target/bg-jobs/sbt_9d8cd5a8/target/309f7c90/1296d6b9/jmh-core-1.37.jar)
[info] WARNING: Please consider reporting this to the maintainers of class org.openjdk.jmh.util.Utils
[info] WARNING: sun.misc.Unsafe::objectFieldOffset will be removed in a future release
[info] # Warmup Iteration   1: 173.684 ops/s
[info] # Warmup Iteration   2: 3469.684 ops/s
[info] # Warmup Iteration   3: 4609.354 ops/s
[info] Iteration   1: 4522.072 ops/s
[info] Iteration   2: 6099.385 ops/s
[info] Iteration   3: 7234.912 ops/s
[info] Result "zio.stm.SinglePermitSemaphoreBenchmark.javaSemaphoreUnfair":
[info]   5952.123 ∩┐╜(99.9%) 24855.288 ops/s [Average]
[info]   (min, avg, max) = (4522.072, 5952.123, 7234.912), stdev = 1362.402
[info]   CI (99.9%): [? 0, 30807.410] (assumes normal distribution)
[info] # JMH version: 1.37
[info] # VM version: JDK 24.0.2, Java HotSpot(TM) 64-Bit Server VM, 24.0.2+12-54
[info] # VM invoker: C:\Program Files\Java\jdk-24\bin\java.exe
[info] # VM options: <none>
[info] # Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
[info] # Warmup: 3 iterations, 1 s each
[info] # Measurement: 3 iterations, 1 s each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Throughput, ops/time
[info] # Benchmark: zio.stm.SinglePermitSemaphoreBenchmark.javaSemaphoreUnfair
[info] # Parameters: (fibers = 2, permits = 1)
[info] # Run progress: 37.50% complete, ETA 00:00:35
[info] # Fork: 1 of 1
[info] WARNING: A terminally deprecated method in sun.misc.Unsafe has been called
[info] WARNING: sun.misc.Unsafe::objectFieldOffset has been called by org.openjdk.jmh.util.Utils (file:/C:/Users/USER/3D%20Objects/zio/target/bg-jobs/sbt_9d8cd5a8/target/309f7c90/1296d6b9/jmh-core-1.37.jar)
[info] WARNING: Please consider reporting this to the maintainers of class org.openjdk.jmh.util.Utils
[info] WARNING: sun.misc.Unsafe::objectFieldOffset will be removed in a future release
[info] # Warmup Iteration   1: 133.762 ops/s
[info] # Warmup Iteration   2: 2681.796 ops/s
[info] # Warmup Iteration   3: 3288.022 ops/s
[info] Iteration   1: 3720.367 ops/s
[info] Iteration   2: 4745.767 ops/s
[info] Iteration   3: 4786.269 ops/s
[info] Result "zio.stm.SinglePermitSemaphoreBenchmark.javaSemaphoreUnfair":
[info]   4417.468 ∩┐╜(99.9%) 11020.056 ops/s [Average]
[info]   (min, avg, max) = (3720.367, 4417.468, 4786.269), stdev = 604.046
[info]   CI (99.9%): [? 0, 15437.524] (assumes normal distribution)
[info] # JMH version: 1.37
[info] # VM version: JDK 24.0.2, Java HotSpot(TM) 64-Bit Server VM, 24.0.2+12-54
[info] # VM invoker: C:\Program Files\Java\jdk-24\bin\java.exe
[info] # VM options: <none>
[info] # Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
[info] # Warmup: 3 iterations, 1 s each
[info] # Measurement: 3 iterations, 1 s each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Throughput, ops/time
[info] # Benchmark: zio.stm.SinglePermitSemaphoreBenchmark.zioSemaphore
[info] # Parameters: (fibers = 1, permits = 1)
[info] # Run progress: 50.00% complete, ETA 00:00:28
[info] # Fork: 1 of 1
[info] WARNING: A terminally deprecated method in sun.misc.Unsafe has been called
[info] WARNING: sun.misc.Unsafe::objectFieldOffset has been called by org.openjdk.jmh.util.Utils (file:/C:/Users/USER/3D%20Objects/zio/target/bg-jobs/sbt_9d8cd5a8/target/309f7c90/1296d6b9/jmh-core-1.37.jar)
[info] WARNING: Please consider reporting this to the maintainers of class org.openjdk.jmh.util.Utils
[info] WARNING: sun.misc.Unsafe::objectFieldOffset will be removed in a future release
[info] # Warmup Iteration   1: 5.887 ops/s
[info] # Warmup Iteration   2: 536.731 ops/s
[info] # Warmup Iteration   3: 692.601 ops/s
[info] Iteration   1: 922.180 ops/s
[info] Iteration   2: 940.330 ops/s
[info] Iteration   3: 1034.144 ops/s
[info] Result "zio.stm.SinglePermitSemaphoreBenchmark.zioSemaphore":
[info]   965.551 ∩┐╜(99.9%) 1096.298 ops/s [Average]
[info]   (min, avg, max) = (922.180, 965.551, 1034.144), stdev = 60.092
[info]   CI (99.9%): [? 0, 2061.849] (assumes normal distribution)
[info] # JMH version: 1.37
[info] # VM version: JDK 24.0.2, Java HotSpot(TM) 64-Bit Server VM, 24.0.2+12-54
[info] # VM invoker: C:\Program Files\Java\jdk-24\bin\java.exe
[info] # VM options: <none>
[info] # Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
[info] # Warmup: 3 iterations, 1 s each
[info] # Measurement: 3 iterations, 1 s each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Throughput, ops/time
[info] # Benchmark: zio.stm.SinglePermitSemaphoreBenchmark.zioSemaphore
[info] # Parameters: (fibers = 2, permits = 1)
[info] # Run progress: 62.50% complete, ETA 00:00:21
[info] # Fork: 1 of 1
[info] WARNING: A terminally deprecated method in sun.misc.Unsafe has been called
[info] WARNING: sun.misc.Unsafe::objectFieldOffset has been called by org.openjdk.jmh.util.Utils (file:/C:/Users/USER/3D%20Objects/zio/target/bg-jobs/sbt_9d8cd5a8/target/309f7c90/1296d6b9/jmh-core-1.37.jar)
[info] WARNING: Please consider reporting this to the maintainers of class org.openjdk.jmh.util.Utils
[info] WARNING: sun.misc.Unsafe::objectFieldOffset will be removed in a future release
[info] # Warmup Iteration   1: 0.879 ops/s
[info] # Warmup Iteration   2: 10.176 ops/s
[info] # Warmup Iteration   3: 159.730 ops/s
[info] Iteration   1: 144.115 ops/s
[info] Iteration   2: 127.206 ops/s
[info] Iteration   3: 122.216 ops/s
[info] Result "zio.stm.SinglePermitSemaphoreBenchmark.zioSemaphore":
[info]   131.179 ∩┐╜(99.9%) 209.390 ops/s [Average]
[info]   (min, avg, max) = (122.216, 131.179, 144.115), stdev = 11.477
[info]   CI (99.9%): [? 0, 340.569] (assumes normal distribution)
[info] # JMH version: 1.37
[info] # VM version: JDK 24.0.2, Java HotSpot(TM) 64-Bit Server VM, 24.0.2+12-54
[info] # VM invoker: C:\Program Files\Java\jdk-24\bin\java.exe
[info] # VM options: <none>
[info] # Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
[info] # Warmup: 3 iterations, 1 s each
[info] # Measurement: 3 iterations, 1 s each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Throughput, ops/time
[info] # Benchmark: zio.stm.SinglePermitSemaphoreBenchmark.zioSemaphore3
[info] # Parameters: (fibers = 1, permits = 1)
[info] # Run progress: 75.00% complete, ETA 00:00:14
[info] # Fork: 1 of 1
[info] WARNING: A terminally deprecated method in sun.misc.Unsafe has been called
[info] WARNING: sun.misc.Unsafe::objectFieldOffset has been called by org.openjdk.jmh.util.Utils (file:/C:/Users/USER/3D%20Objects/zio/target/bg-jobs/sbt_9d8cd5a8/target/309f7c90/1296d6b9/jmh-core-1.37.jar)
[info] WARNING: Please consider reporting this to the maintainers of class org.openjdk.jmh.util.Utils
[info] WARNING: sun.misc.Unsafe::objectFieldOffset will be removed in a future release
[info] # Warmup Iteration   1: 17.952 ops/s
[info] # Warmup Iteration   2: 1007.060 ops/s
[info] # Warmup Iteration   3: 1355.255 ops/s
[info] Iteration   1: 1210.174 ops/s
[info] Iteration   2: 1659.662 ops/s
[info] Iteration   3: 1878.727 ops/s
[info] Result "zio.stm.SinglePermitSemaphoreBenchmark.zioSemaphore3":
[info]   1582.855 ∩┐╜(99.9%) 6218.015 ops/s [Average]
[info]   (min, avg, max) = (1210.174, 1582.855, 1878.727), stdev = 340.830
[info]   CI (99.9%): [? 0, 7800.869] (assumes normal distribution)
[info] # JMH version: 1.37
[info] # VM version: JDK 24.0.2, Java HotSpot(TM) 64-Bit Server VM, 24.0.2+12-54
[info] # VM invoker: C:\Program Files\Java\jdk-24\bin\java.exe
[info] # VM options: <none>
[info] # Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
[info] # Warmup: 3 iterations, 1 s each
[info] # Measurement: 3 iterations, 1 s each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Throughput, ops/time
[info] # Benchmark: zio.stm.SinglePermitSemaphoreBenchmark.zioSemaphore3